Move from package snic to unsafe
