+++ stensorflow/src/main/scala/org/ekrich/tensorflow/unsafe/tensorflow.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ekrich.tensorflow.snic
+package org.ekrich.tensorflow.unsafe
 
 import scalanative.unsafe._
